+++ collector/collector.go
@@ -14,6 +14,9 @@ const (
 	namespace = "mysql"
 	// Math constant for picoseconds to seconds.
 	picoSeconds = 1e12
+	// Query to check whether user/table/client stats are enabled.
+	userstatCheckQuery = `SHOW VARIABLES WHERE Variable_Name='userstat'
+		OR Variable_Name='userstat_running'`
 )
 
 var logRE = regexp.MustCompile(`.+\.(\d+)$`)
@@ -30,6 +33,17 @@ func parseStatus(data sql.RawBytes) (float64, bool) {
 		return 1, true
 	}
 	if bytes.Compare(data, []byte("No")) == 0 || bytes.Compare(data, []byte("OFF")) == 0 {
+		return 0, true
+	}
+	// SHOW SLAVE STATUS Slave_IO_Running can return "Connecting" which is a non-running state.
+	if bytes.Compare(data, []byte("Connecting")) == 0 {
+		return 0, true
+	}
+	// SHOW GLOBAL STATUS like 'wsrep_cluster_status' can return "Primary" or "Non-Primary"/"Disconnected"
+	if bytes.Compare(data, []byte("Primary")) == 0 {
+		return 1, true
+	}
+	if bytes.Compare(data, []byte("Non-Primary")) == 0 || bytes.Compare(data, []byte("Disconnected")) == 0 {
 		return 0, true
 	}
 	if logNum := logRE.Find(data); logNum != nil {