+++ collector/global_variables.go
@@ -4,6 +4,8 @@ package collector
 
 import (
 	"database/sql"
+	"regexp"
+	"strconv"
 	"strings"
 
 	"github.com/prometheus/client_golang/prometheus"
@@ -26,10 +28,12 @@ func ScrapeGlobalVariables(db *sql.DB, ch chan<- prometheus.Metric) error {
 
 	var key string
 	var val sql.RawBytes
-	var mysqlVersion = map[string]string{
-		"innodb_version":  "",
-		"version":         "",
-		"version_comment": "",
+	var textItems = map[string]string{
+		"innodb_version":         "",
+		"version":                "",
+		"version_comment":        "",
+		"wsrep_cluster_name":     "",
+		"wsrep_provider_options": "",
 	}
 
 	for globalVariablesRows.Next() {
@@ -44,15 +48,55 @@ func ScrapeGlobalVariables(db *sql.DB, ch chan<- prometheus.Metric) error {
 				floatVal,
 			)
 			continue
-		} else if _, ok := mysqlVersion[key]; ok {
-			mysqlVersion[key] = string(val)
+		} else if _, ok := textItems[key]; ok {
+			textItems[key] = string(val)
 		}
 	}
-	// Create mysql_version_info metric
+
+	// mysql_version_info metric.
 	ch <- prometheus.MustNewConstMetric(
 		prometheus.NewDesc(prometheus.BuildFQName(namespace, "version", "info"), "MySQL version and distribution.",
 			[]string{"innodb_version", "version", "version_comment"}, nil),
-		prometheus.GaugeValue, 1, mysqlVersion["innodb_version"], mysqlVersion["version"], mysqlVersion["version_comment"],
+		prometheus.GaugeValue, 1, textItems["innodb_version"], textItems["version"], textItems["version_comment"],
 	)
+
+	// mysql_galera_variables_info metric.
+	if textItems["wsrep_cluster_name"] != "" {
+		ch <- prometheus.MustNewConstMetric(
+			prometheus.NewDesc(prometheus.BuildFQName(namespace, "galera", "variables_info"), "PXC/Galera variables information.",
+				[]string{"wsrep_cluster_name"}, nil),
+			prometheus.GaugeValue, 1, textItems["wsrep_cluster_name"],
+		)
+	}
+
+	// mysql_galera_gcache_size_bytes metric.
+	if textItems["wsrep_provider_options"] != "" {
+		ch <- prometheus.MustNewConstMetric(
+			newDesc("galera", "gcache_size_bytes", "PXC/Galera gcache size."),
+			prometheus.GaugeValue,
+			parseWsrepProviderOptions(textItems["wsrep_provider_options"]),
+		)
+	}
+
 	return nil
+}
+
+// parseWsrepProviderOptions parse wsrep_provider_options to get gcache.size in bytes.
+func parseWsrepProviderOptions(opts string) float64 {
+	var val float64
+	r, _ := regexp.Compile(`gcache.size = (\d+)([MG]?);`)
+	data := r.FindStringSubmatch(opts)
+	if data == nil {
+		return 0
+	}
+
+	val, _ = strconv.ParseFloat(data[1], 64)
+	switch data[2] {
+	case "M":
+		val = val * 1024 * 1024
+	case "G":
+		val = val * 1024 * 1024 * 1024
+	}
+
+	return val
 }