+++ collector/info_schema_userstats.go
@@ -8,6 +8,7 @@ import (
 	"strings"
 
 	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/common/log"
 )
 
 const userStatQuery = `SELECT * FROM information_schema.user_statistics`
@@ -50,6 +51,22 @@ var (
 		"BINLOG_BYTES_WRITTEN": {prometheus.CounterValue,
 			prometheus.NewDesc(prometheus.BuildFQName(namespace, informationSchema, "user_statistics_binlog_bytes_written_total"),
 				"The number of bytes written to the binary log from this user’s connections.",
+				[]string{"user"}, nil)},
+		"ROWS_READ": {prometheus.CounterValue,
+			prometheus.NewDesc(prometheus.BuildFQName(namespace, informationSchema, "user_statistics_rows_read_total"),
+				"The number of rows read by this user's connections.",
+				[]string{"user"}, nil)},
+		"ROWS_SENT": {prometheus.CounterValue,
+			prometheus.NewDesc(prometheus.BuildFQName(namespace, informationSchema, "user_statistics_rows_sent_total"),
+				"The number of rows sent by this user's connections.",
+				[]string{"user"}, nil)},
+		"ROWS_DELETED": {prometheus.CounterValue,
+			prometheus.NewDesc(prometheus.BuildFQName(namespace, informationSchema, "user_statistics_rows_deleted_total"),
+				"The number of rows deleted by this user's connections.",
+				[]string{"user"}, nil)},
+		"ROWS_INSERTED": {prometheus.CounterValue,
+			prometheus.NewDesc(prometheus.BuildFQName(namespace, informationSchema, "user_statistics_rows_inserted_total"),
+				"The number of rows inserted by this user's connections.",
 				[]string{"user"}, nil)},
 		"ROWS_FETCHED": {prometheus.CounterValue,
 			prometheus.NewDesc(prometheus.BuildFQName(namespace, informationSchema, "user_statistics_rows_fetched_total"),
@@ -108,6 +125,17 @@ var (
 
 // ScrapeUserStat collects from `information_schema.user_statistics`.
 func ScrapeUserStat(db *sql.DB, ch chan<- prometheus.Metric) error {
+	var varName, varVal string
+	err := db.QueryRow(userstatCheckQuery).Scan(&varName, &varVal)
+	if err != nil {
+		log.Debugln("Detailed user stats are not available.")
+		return nil
+	}
+	if varVal == "OFF" {
+		log.Debugf("MySQL @@%s is OFF.", varName)
+		return nil
+	}
+
 	informationSchemaUserStatisticsRows, err := db.Query(userStatQuery)
 	if err != nil {
 		return err